Renames countLinks to getLinkStatistics
