+++ js/analyses/getLinkStatistics.js
